refactor: extract QueueList from queue-panel into shared module
Move QueueList (and its NowPlaying helper + TOUCH_SENSOR_OPTIONS) out of
queue-panel.tsx into a new queue-list.tsx so it can be shared between the
audio player popover/sheet and the dashboard queue section. The list now
takes an optional maxHeight prop — QueuePanel passes "50vh" to keep the
popover/sheet scroll cap; future dashboard usage will omit it so the list
flows inside its card.

QueuePanel public behaviour (popover + sheet branches, click-to-play,
remove, clear-all, DnD wiring) is unchanged.

Refs #394.
